fix: clear observation context audit drift
